Update drop dates for revised drop style
Wizards has effectively reverted a small part of the rotation change
from 2014 -- rather than the first set of every block dropping the
oldest block in Standard, the first set of every *other* block will drop
the oldest *two* blocks in Standard, returning us to the "four sets
leave simultaneously every two years" model from pre-2014.

http://magic.wizards.com/en/articles/archive/news/revisiting-standard-rotation-2016-10-19

This change updates drop dates on the website and in the API to comply,
and revises the description of recent updates on the website.

Closes #35