Design update, especially mobile
Big update to the layout with the goal of being more mobile friendly,
and a side goal of feeling more like an app (because we are a PWA now!).

While I was at it I tried to fix #153 from two fronts:
1. For future sets, background is absent, border is absent, and the set
icon, while still absent as before, feels more absent because set icons
are much more prominent now;
2. For past sets, actually decided to keep the old layout just to have a
more striking visual difference.

In addition to the set design changes this also adds a navbar to the
bottom for mobile-sized screens.

Also, we are on the Play Store now!
https://play.google.com/store/apps/details?id=xyz.appmaker.jamscv