Mess w/ display re: more sets in Standard at once
This is a throwback to the original display we used to have using list
groups. I think we moved away from it due to small issues here and small
issues there, but overall the design ended up blocky and not very fluid.
Gonna try to start again from here, and Standard drastically changing
its min and max sizes is a good time to make the change.